backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Firebase & Google Cloud
firebase-admin==6.3.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-cloud-aiplatform==1.38.0


firebase-admin>=6.0.0


# API & Validation
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
marshmallow==3.20.1

# ML / AI
# torch==2.2.0
# torchvision==0.17.2
# tensorflow==2.20.0
# # CHANGE ↓ tensorflow-lite package name and version
# # tflite-runtime==2.20.0
# scikit-learn==1.3.2
# Pillow==10.1.0
# opencv-python==4.8.1.78
# numpy==1.26.4
# pandas==2.1.4

torch>=2.2.2,<3.0
torchvision>=0.17.2,<1.0
tensorflow>=2.16.0,<3.0
scikit-learn>=1.4.0
numpy>=1.26.0,<2.0.0
pandas>=2.2.0,<3.0
opencv-python>=4.9.0
Pillow>=10.3.0

# MQTT & IoT
paho-mqtt==1.6.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Task Queue
celery==5.3.4
redis==5.0.1

# PDF Generation / Visualization
reportlab==4.0.7
matplotlib==3.8.2
plotly==5.18.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
